- Remove BIT pcode instruction.  It is not used.  It was once used for non-standard check if a bit ws set in an integer. - Add another simple test file. - Reformat SETOP instruction.  An 8-bit argument is sufficient.

+++ insn16/prun/psetops.c
@@ -377,7 +377,7 @@ static uint16_t pexec_BitsInNibble(uint8_t nibble)
  *
  ****************************************************************************/
 
-int pexec_setops(struct pexec_s *st, uint16_t subfunc)
+int pexec_setops(struct pexec_s *st, uint8_t subfunc)
 {
   const uint16_t *src1;
   const uint16_t *src2;
